tests: Start the hosted mode server on a random port
Preparation for running e2e tests in parallel.

The tests no longer rely on the hosted mode server running on a
specific port. Each time we run the tests, we pick a new port at
random from a range and try to start the hosted mode server with
it. We retry up to a certain threshold if the port was in use,
but we never retry with the same port.

When we have parallel test runners that don't know about each
other, this will ensure that they can all start their own server
successfully, as long as retries > the number of parallel
runners.

This will also make the parallel runners more robust to ports
that are taken, but not by mocha. We aren't sure what ports
will be available on the bots so this adds some redundancy
there.

Bug: 1101784
Change-Id: Ia846eeef027103cf735dcc9a690cd7adc630b382