[issues] Sort issues alphabetically
This CL sorts issues alphabetically by issue title. Including other
sort orders, like a lexical sorting first considering a category and
is now easy to do, should we decide to do so.

The sorting uses logic already present in the TreeElement: There a
comparator can be supplied, that is applied on all levels of the
tree. This CL introduces a common base class OrderedTreeElement for
all TreeElements used in the issues view, and initializes a string
used for sorting accordingly. This means that we need to introduce
some constants to get the sorting of issue body / affected resources
and link right.

Bug: chromium:1072335
Change-Id: Id563d17ebc458434d72ab706d013d758dbd7ea73