tests: Request a free port for the hosted mode server
Rather than hardcode the port to use, use port 0 which
requests any free port from the OS. We report the port
we end up using to the caller over IPC, if an IPC
channel was opened by the caller.

We wait for the port to get reported before starting
the chrome instance.

Previously we started the hosted mode server and then
started chrome without waiting for it to start. This is
probably fine in most situations because chrome takes a
while to start and we need to do extra setup there, so
we have a long time before the server needs to respond.

For parallel test mode, we need to make sure we start
the server on a free port. This also makes the mocha
runner more robust to failures e.g. if we accidentally
leave the server running and the port is taken.

This doesn't slow down the startup time of the e2e
test runner at all, according to timing of local runs.

Bug: 1101784
Change-Id: Ibb14d5a177ba5f9e257743824770bff3d5ccb033