Force inclusion of front_end/formatter_worker for swarm bots
The swarm bots use :devtools_frontend_resources to figure out
which files to send to the bots. Even though we have specified the
generated TypeScript files and included in the GRD file, if the
file does not exist on the filesystem on the bots.

Therefore, if we force add it to the dependencies, we can make sure
that the file is forcefully added.

This is not the proper longterm fix, as it should find all dependencies
recursively in the GN hierarchy tree itself. However, since we are
using `deps` rather than `public_deps` in our actions, that is a
plausible explanation as to why the swarm bots are not finding the files.

Once we are back on stable ground, we are going to test that hypothesis
and prevent this issue from occurring again.

TBR=aerotwist@chromium.org

Change-Id: I0cb00db954389ec1fbc03a1990d7fd1712c1ac51