Move front_end html_entrypoints to data_deps
Based on the explanation by dpranke@ provided in
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2292011/4#message-9a4e191b1512321db4018a327ece20e87e9ef30c
we were able to figure out that we need to specify the entrypoints in
`front_end/` as `data_deps`.

The difference between `public_deps` and `data_deps` is that
`public_deps` only enforces compiler-order task building. When the CQ
bots figure out what files are required to be sent to swarming, they
instead traverse `data_deps` and `data` attributes on actions.

Before this change, CQ sent `integration_test_runner.html` to the bots,
because it was part of the `data_deps` of
`:devtools_frontend_resources_data`. Now, we moved it to `data_deps` of
`front_end`, because `front_end` is one of the `data_deps` of
`:devtools_frontend_resources_data` as well.

Follow-up changes will fix `devtools_{module,entrypoint}` to correctly
specify the `data_deps` and `data` hierarchies, to make sure that all
files that DevTools builds will be sent to CQ.

Change-Id: Id9e79eabd6e57da30a8f0d1783b1f63ff2fd7eef