Re-implement 'Reload required' for some settings
Previously, we had implemented a "Reload required" option for theme
changes:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2107500

However, this change was reverted for stability. In consideration
of the ways in which not only the themes are affected, but also
experiments, and how the theme might cause a change in the event
that the OS-preferred theme changes (see
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2279841
), this change presents an Infobar prompt to reload DevTools in
case such a setting change is required.

After: https://imgur.com/a/4SdnEYC
Also after: https://imgur.com/a/3YuHOP5 (Command palette)

Bug: chromium:1001549, chromium:1100827
Change-Id: Ieb0a667c2b17f75159a11faa30327c7a097d9614