Change for TabbedPane title trimming
As noted in CRBug 1105480, the Sources tab trims the aria-label
of the tab because the component creating the tabs unconditionally
trims the text based on its length. This approach fixes this issue
by instructing the TabbedPane itself to handle title trimming to
a given maximum length, and trimming the text within the
implementation of TabbedPaneTab.

Before/After: https://imgur.com/a/CG4zAip

Bug: 1105480
Change-Id: I6d345e52c31398c368473151f897f9f32101b7e2