Ship new ElementsBreadcrumbs to production
This CL ships the new elements breadcrumbs component to production,
replacing the old legacy breadcrumbs with the new custom element that
we've built to replace it.

Note that I haven't renamed the file from `NewElementsBreadcrumbs`. I wanted to
keep the diff on this CL clear and easy to follow. I plan to follow this up
with another CL that does the renaming.

Change-Id: I5c3df09456101ce03072f3eeb27f6a1c568cf9c7