Upgrade TypeScript ESLint dependencies.
DISABLE_THIRD_PARTY_CHECK=node_modules update

This was an attempt to fix https://crbug.com/1102821 however the issue still
appears. I expect it will until TS4.0 is out of beta and the ecosystem catches
up.

We were however on a much older version so I took the opportunity to update us
to the current latest version.

Change-Id: I8332409e9de50d60f6284eef5098b55a7f9535c4