Move acorn entrypoint to front_end/third_party/acorn
Instead of including the acorn entrypoint in `formatter_worker`,
we move it to the same directory as `acorn` lives in.

Note that we still bundle all extensions of acorn in the Acorn
bundle itself, as such we have to explicitly include only
`third_party/acorn/acorn.js` from the imports in Rollup.

R=aerotwist@chromium.org,bmeurer@chromium.org

Change-Id: Ic7b855fba38e05edd19b7c0fc352d81c00a32497