Grid: Different color for grid extended lines
We have decided to show the extended lines by default in the v2 grid
overlay, coming soon.
Right now those line extensions have the same color and opacity as the
grid lines themselves, and I believe that makes the grid a little bit
harder to see.
The design spec also defines different color/opacity for the extensions.

To do so, we'd normally have to introduce a new property to use when
asking the Overlay domain to highlight the grid, and we can do this.
But as a first step, I thought I'd reuse the same default color that's
defined for ruler extension lines already. This way it's a really small
change and it's consistent with these other lines.
If we do want specific colors for these grid extension lines, then I can
update the CL.

Bug: 1047356
Change-Id: I9f640c6ad8437a39961a4fb720dac9f015c08964