Grid: Handle separate row/col color and dash and highlight on grid-template hover
This is based on the following backend change:
https://chromium-review.googlesource.com/c/chromium/src/+/2302230/

The feature that this change allows is for users to hover over
grid-template-rows and grid-template-columns. When doing so, the rows and
columns get highlighted in the page, accordingly.

The backend CL emits separate information for rows and cols to make this
possible, so this frontend change contains overlay updates to deal with
this.
This frontend change also contains the necessary code to do the on-hover
highlight.

2 ideas for follow-up changes:
* Only showing the explicitly defined lines when hovering
  over grid-template-rows/columns, since those are the ones defined by
  these properties.
* Show the track sizing labels and line name labels when those do land
  since these are also defined by grid-template-rows/columns properties.

Change-Id: I0fed00f16a060fcba303a25e2d47d2d7df06bf5c