Grid: Adding a new setting and telemetry for the line name overlay feature
This adds a new option in the grid settings to show line names.
The option is added to the line label drop-down, so that it is possible to show
- no labels
- positive numbers
- negative numbers
- both numbers
- names

Screenshot: https://imgur.com/APgkZbS

We're keeping the same telemetry already in place for line numbers, just
adding a new option to it.

Bug: 1100385
Change-Id: I36d150b003ffcda1ecf122ff97897fc318f82819