Improve bridges generation interface detection
This CL fixes the component bridges so that it can find nested interfaces. That
is, if it sees:

```
interface Foo {
  bar: Bar
}
```

Before this CL only the `Foo` interface would end up in the _bridge.js as a
Closure typedef. Now, with this CL, it will also include the `Bar` interface in
the _bridge.js file too.

Change-Id: I87a614fd5620c6572a062458a3e4b7dd30ca86e6