Fix bad border colour on computed styles
Because the border declaration was missing a defined colour it was not
inverted correctly in dark mode, meaning the border would be black which
is not easily spotted when in dark mode. To fix this I picked out the colour that was being used in light mode and explicitly defined it.

By declaring a specific black colour, it is then inverted into white in
dark mode and it's much clearer as the screenshots below show (no light mode included as this change is a no-op in light mode).

Before, dark mode (note the barely visible black border around the
margin): https://imgur.com/OuaagvW

After, dark mode: https://imgur.com/ecntqHp

Bug: 1106683
Change-Id: I2a834d23f964efaa6757636460b8b5136bfdfb23