Fix frame tree persistence of selected item
When refreshing the page with DevTools opened, when refreshing
DevTools or when navigating back to a previously opened page, the frame
tree in the application panel should pre-select the element which was
last selected on the previous visit.

This works by comparing itemURLs of the tree elements with previously
persisted itemURLs. In the previous implementation any match between
those 2 arrays was considered a match. This could lead to the
selection jumping around on page refresh because there were multiple
matches. With this CL a match needs to start at the root node and
progresses along parent-child relationships.

Bug: https://crbug.com/1101262
Change-Id: Ica7275be6aaf13df5de1e800e9fbabf1449ec8bd