[DevTools] Fix of crbug 1107564
A missing check in json parsing caused loading of some of the standard
devices to fail, which leads to the bug: after user selects additional
devices, the device list does not appear to persisted - they persisted
but the reading failed, the "show" flag became Default, causing no show.

Tests to be added later.

Bug: 1107564
Change-Id: Idea0927ad108e2467f667329df8a70040696f9df