WIP: Removing "other origins" UI from DevTools Service Worker tab
In this CL, we are removing the drop down menu from DevTools that shows
the service worker registrations from other origins which would allow
us to remove the 'problematic' function GetAllRegistrations() which was
causing the crashes.

We are linking to the chrome://serviceworker-internals page in a
footnote for the occasion that a developer may want to see service
workers from other origins. More details about the plan can be found
at the bottom of the document here:
https://docs.google.com/document/d/1DdGq832j1ShNZosFOHQ72qVd-SeYeRgqypbz7nIhz5w/edit#heading=h.dsnztm8lg04x

Bug: 807440
Change-Id: I63e07e95c5e6cde1085227c15b8fc6b915dbb908