Unify panelShown telemetry logic
Previously `panelShown()` had different logic for logging telemetry if a panel was shown in the main view or if a panel was shown in the drawer. It prepended the word ‘drawer-‘ to the panel id before using that as a key to find the enum value to log in the histogram.

With the introduction if movable tabs it no longer makes sense to have this distinction, since a panel location can be changed by the user. After a panel is moved, the logic can’t find the correct key in the `PanelCodes` enum.

To fix this, we can stop prepending the ‘drawer-‘ string and just use the panel id. What we send to the backend is the corresponding number value of the enum, so there is no risk of changing the keys. I think on the backend we could keep the enum labels the same for historical consistency (I don’t feel too strong about this) because enum values are the same.


Bug: 1075732
Change-Id: I3f356605aa263fef2d66a0bac90f39362d764feb