[bandaid] Let issue banner scroll out of view in console
It has been reported by users through multiple channels that the
"issues detected" message is obtrusive and disrupting the debugging
workflow. This is a back-mergable fix that lets the "issues detected"
message scroll out of view. This shoud fix the behavior for M85,
and opens the way for reworking this for M86.

Screenshot: https://imgur.com/a/IkvGFAi
Bug: chromium:1096481
Change-Id: I6014452bb06ca5ccb349ce2abcf51b918bd345c3