Fix interface detection in union types in component bridges generator
If a setter took an object literal where the value was a union type
(e.g. `foo: Foo|null`) we would ignore it, and as such the bridge generator
would not include `Foo` in the bridge. This CL adds a check for union types
where we loop over each member and check if there's an interface in there.

Change-Id: I23039dea3af12d1bd0f3a3382e320a8c053cacc1