Use filter value autocompletes for backdrop-filter
Currently the `backdrop-filter` property's autocomplete values are all
color values. This CL changes these autocompletes to the same as the
`filter` property.

This issue also shows the underlying problem of using many manually
kept preset lists for CSS property name and value autocompletes, and
this underlying problem will be fixed separately, because it requires
major refactors on how we autogenerate properties and autocompletes.

Bug: chromium:1014289
Change-Id: I3e936dd554f39c35fdcfe982859e4e12d94193f1