Add support for `type X =` to component bridges
We previously only looked for interfaces, but sometimes we use types, so
now we support both.

There's still features that aren't supported - e.g. if you extend a
type, that won't work, but that's next on the list. Similarly we don't
parse union types to see if they contain other types, but that will also
be done in another follow-up CL. It's easier to incrementally add these
features than do them in one big go.

Change-Id: Ifa9c647b4b261a1ea1ca237d554733fed23221da