Component bridges: support types that extend types
With this change we support:

```
type Person = {}

type SomeOtherPerson = Person & {...}
```

Closure doesn't have a concept of types being extended, so the bridges
code will instead parse all the types and generate a new type for
Closure made up of all the members of any types that are extended.

Note that this CL does _not_ add support for interfaces that extend
another, ONLY types. Interface extending is coming in a follow up CL.

Change-Id: I3f8253663430d16f28f1bdc3c40e582d8ff09248