[bandaid] Let issue banner scroll out of view in console
It has been reported by users through multiple channels that the
"issues detected" message is obtrusive and disrupting the debugging
workflow. This is a back-mergable fix that lets the "issues detected"
message scroll out of view. This shoud fix the behavior for M85,
and opens the way for reworking this for M86.

This additionally fixes the ugly alignment of the "Go to issues" button,
which causes the issues banner to become especially large if DevTools
is opened on small widths.

Screenshot: https://imgur.com/a/ny69asC
Bug: chromium:1096481
Change-Id: I6014452bb06ca5ccb349ce2abcf51b918bd345c3