Include nested interfaces when extending types
This CL fixes a bug in the bridges generation where nested types would
not correctly be added to the bridge when they were referenced via an
extended type.

For example, consider this code:

```
interface Detail {
  id: number;
}

type NamedThing = {
  name: string;
}

type Person = NamedThing & { details: Detail[] };
```

The bridges generator will correctly recognise that it needs to define a
`Person` typedef that includes all the members of `NamedThing` and also
the `details` field. But without this CL it will not realise that the
`Detail` interface is also referenced and therefore needs to be added to
the bridge.

This CL ensures when we extend types that we check their members for any
interfaces that are also required.

Change-Id: I5cf0de3ebc63dd56e71af8284ba6fe9fc10b396e