Fix invisible border on computed styles.
A fix I applied to try to improve the constrast of the "border" text
lead to it being invisible in dark mode when hovering over the padding
part of the computed styles.

Removing this fix fixes the attached bug but does decrease the constrast
of the border text when hovering over it, but I think that's a
reasonable trade off. Really it would be great to pick better colours
for the computed styles pane so we don't have to try to hack our way to
good contrast.

Fixed: 1110271
Change-Id: I5c991cc2507cbc0aadb655b5d246581afd4cc26d