Adds global list of files expected for the GRD
We've found that files missed from the GRD cause issues in Chromium,
since it looks to the GRD/pak in order to serve the DevTools frontend.
Standalone builds of the DevTools frontend (as well as using
--custom-devtools-frontend in Chromium) do not present with the issue,
since in those cases the files are loaded directly from disk (rather
than via the GRD/pak file).

This CL introduces two new lists: a global list of files required for
the Release build (is_debug=false) of Chromium, and a second list of
files required for the Debug build (is_debug=true). These lists are
cross-referenced to ensure no overlap, thus in Release we use the first
list, and in Debug we use both.

In the root BUILD.gn we now check that files that are inputted to the
GRD builder action map exactly to the list of files expected for the
GRD, and any mismatch causes a build failure.

R=danno@chromium.org

Change-Id: I5e215eb5dc249123ab07210baa3eeb00da225410