HeapSnapshot: Process detached'ness state
This change allows for pcoessing detached'ness of an object in the
frontend as follows:
- Nodes are either detached are attached.
- Node state is propagated through the graph:
  1. Any object reachable from an attached object is itself attached.
  2. Any object reachable from a detached object that is not already
     attached is considered detached.
- Node name is adjusted with a "Detached " prefix.

This CL prepares for a change where detached'ness is not propgated
forward during snapshot creation anymore. In future, V8 will create the
snapshot internally for C++ objects. Blink-speciufic information is
gathered using embedder callbacks from V8 which is then processed on the
frontend.

Bug: chromium:1110816
Change-Id: I5356dc61abeabf6d99940ef2003d5763a9324832