Fix errors with `Object` type in component bridges
The component bridge code would think `Object` is a type that it needs
to define as a Closure typedef but in fact it doesn't, it's built-in to
TS and maps to Closure's Object type. So if we find any, don't try to
define them to be converted, and instead output it directly as an
`Object` in Closure.

Change-Id: Ied707d131b95a8f595832f6e2080722f49522efb