Grid: Set default CSS grid setting to extended dashed lines
This CL changes the default setting for grid lines to be extended
dashed based on the recent user study.

Note that that this will only change the default for users
who enable the experiment for the first time after this change, as
otherwise it will respect existing settings.

Also note that extended lines are still subject to: crbug.com/1106681,
as the extended lines do not rotate with grids (this is also the
behavior of the inspector with the "show rulers" setting enabled)

Bug: 1109215
Change-Id: I5b6746d0f0ea78cc134caa60c465f0155aa0168e