CSS Grid: add background to row/column gaps for default highlighting
This CL changes our inspector overlay to always place a different colored
background behind row/column gaps in addition to the existing hatch
shading. This is in response to a recent user study that indicated
a desire for more contrast between the highlight for the content box (blue)
and the grid gaps (previously only hatch filled).

Before:

After:

Bug: 1109213
Change-Id: I16f11e2c6aad276014cc7771342807c7ebaa6ecc