Use event to trigger Network Panel reveal & filter
There is a circular dependency between the Cookie Table and Network
Panel. The dependency is strongest from the Network Panel to the Cookie
Table, since it uses the Cookie Table to render cookies in its view.
The Cookie Table dependency, however, is relatively weak, used only to
reveal cookie requests in the Network Panel. This circular dependency
blocks the migration to devtools_entrypoint since gn now knows about it
and does not support it.

This CL changes the behavior of the Cookie Table to dispatch an event
with the required filter information, which is picked up by the Network
Panel. This decouples the two panels, and allows us to break the
circular dependency.

R=jacktfranklin@chromium.org

Change-Id: I933c281c6755486e3dc14bbd852e8422b82b21a9