Align bridges generator with Clang format
The bridges were outputting functions as:

`x() {}`

But Clang would change that to:

```
x() {
}
```

It was a bit annoying that everytime the bridge file changed the
presubmit would fail as it reformatted, so this change brings the
bridges fully in line with Clang (e.g. a presubmit shouldn't make
changes to it).

Change-Id: Ie27a73b3d2933cf967c320d3c7c6122e4d6d806d