Generate COEP issues based on issues generated in the back-end
This also undoes a change where certain frame navigations are
ignored in the front-end to improve debuggability. The reason is
that the change doesn't work for OOPIFs, and as those are very
common with COEP issues, prevents debugability there. We likely
need to adopt a fix similar to https://crrev.com/c/2144163 to
solve this issue.

Screenshot: https://imgur.com/a/zfeMiHl
Bug: chromium:1051466, chromium:1078823
Change-Id: I943ba9a38909b6f4505d24bdd89263f9e58cf730