Style issue bar to mimic regular console message
In order to make the issues banner less intrusive, this CL changes its
style so that it looks like a regular console message. In oder to
conserve the ability to potentially back-merge this change, code
changes are kept to a minimum resorting to mostly CSS changes.

For Chrome M87 it is planned to move this functionality entirely to
a console message, so that it does not necessarily always show up as
the first console message, but is part of the console message flow
instead.

Video: https://i.imgur.com/zM81aEC.mp4

Bug: chromium:1096481, chromium:1068116

Change-Id: I0e6b87fbe56c20edb409dc898a1542a88995041e