[TypeScript] Ensure no hardlinks before tsc runs
When TypeScript writes to pre-existing files, it will overwrite the
contents, but it won't create a new file per se. If files in gen/ were
previously created by devtools_pre_built they will be hardlinked to the
original source file, thus any changes tsc makes to the file in gen will
be reflected back to the source. This causes an issue with ninja, since
it believes on the next run that the source file has changed.

This CL checks the files that TypeScript is about to create, by
walking the target out directory, and checking the corresponding file in
the front_end. If the two files have the same inode, they are hardlinked
and the bug will likely manifest. As such the script will remove the
copy of the file in gen on the assumption that the TypeScript compiler
will be writing it as part of its process.

Change-Id: I11a23fce764101eb237e434a64159223ef8d700e